chore(feedback): implement feedback refactor routes file structure to one file rename class to comform to convention
[finishes #165377693]
