+++ server/middleware/validation/userValidation.js
@@ -1,0 +1,121 @@
+import Validate from '../../helpers/validation';
+import Users from '../../models/users';
+
+class UserValidation {
+  /**
+   *helps validate that the data passed in through the signp end point is as expected
+   *
+   * @static validateSignup
+   * @param {object} req
+   * @param {object} res
+   * @param {function} next
+   * @returns error message that guides for correct use of this endpoint
+   * @memberof UserValidation
+   */
+  static validateSignup(req, res, next) {
+    const {
+      firstName, lastName, email, phoneNumber, password,
+    } = req.body;
+
+    const requiredNotGiven = Validate.requiredFieldIsGiven({
+      firstName,
+      lastName,
+      email,
+      phoneNumber,
+      password,
+    });
+
+    if (requiredNotGiven) {
+      return res.status(400).json({
+        status: 400,
+        error: requiredNotGiven,
+      });
+    }
+
+    const isStringNotValid = Validate.stringType({
+      firstName,
+      lastName,
+      password,
+    });
+
+    if (isStringNotValid) {
+      return res.status(400).json({
+        status: 400,
+        error: isStringNotValid,
+      });
+    }
+
+    const isEmailNotValid = Validate.emailType({ email });
+    if (isEmailNotValid) {
+      return res.status(400).json({
+        status: 400,
+        error: isEmailNotValid,
+      });
+    }
+
+    const isPasswordSecure = Validate.minPasswordLength({ password });
+    if (isPasswordSecure) {
+      return res.status(400).json({
+        status: 400,
+        error: isPasswordSecure,
+      });
+    }
+
+    const isPhoneNumberNotValid = Validate.phoneNumberValid({ phoneNumber });
+    if (isPhoneNumberNotValid) {
+      return res.status(400).json({
+        status: 400,
+        error: isPhoneNumberNotValid,
+      });
+    }
+
+    if (!Validate.emailIsUnique(email)) {
+      return res.status(400).json({
+        status: 400,
+        error: `Account with email: ${email} already exists`,
+      });
+    }
+
+    next();
+  }
+
+  /**
+   *validates that the expected data are passed to the signin route
+   *
+   * @static validateSignup
+   * @param {object} req
+   * @param {object} res
+   * @param {function} next
+   * @returns error messages to guide users in passing expected input to the signin endpoint
+   * @memberof UserValidation
+   */
+  static validateSignin(req, res, next) {
+    const { email, password } = req.body;
+
+    const userDetails = Users.find(user => user.email === email);
+    const requiredNotGiven = Validate.requiredFieldIsGiven({ email, password });
+
+    if (requiredNotGiven) {
+      return res.status(400).json({
+        status: 400,
+        error: requiredNotGiven,
+      });
+    }
+
+    if (!userDetails) {
+      return res.status(401).json({
+        status: 401,
+        error: 'Email or password is wrong',
+      });
+    }
+
+    if (password !== userDetails.password) {
+      return res.status(401).json({
+        status: 401,
+        error: 'Email or password is wrong',
+      });
+    }
+    next();
+  }
+}
+export default UserValidation;
